Added examples to prove that the multi-component modelling is fully functional
Added examples to prove that the multi-component modelling is fully functional. Also, added the ability to simulate 2-component gaussian data in the simulate_cosinor function. At the moment, this is just for demonstration but I plan to integrate it more fully into the script. For example, at the moment the period and component-changes are fixed, but it would be useful if the user could specify these.