Fixed the no-intercept specification
Adjusted the way the amp.acro function creates the formula so that it respects a no-intercept specification. The user can now also specify a mesor for only one group if they wish

Started documenting SRR standards
Adjusted update_covnames function again. Previously, this function would not work if there are no mesor covariates. The changes make this code as flexible as possible