Add multi-component polar plotting
Currently, there is much wrong with this script. I spent a the whole day trying to rewrite the script using coords_polar(), but this was incredibly inconsistent. At the moment, I've reverted the current script to a more basic form, but there is still much to do with it. Firstly, the ellipses aren't oriented properly, and secondly, the center points of the components are not updating correctly.