removed a negative sign
Our acrophase values were actually estimating the time for the trough, not the peak to occur due to an inappropriate negative sign in amp_acro. Interestingly, this didn't affect the plots because the negative sign was only used to find the transformed estimates, and the plots were genereted using the raw estimates. Now, acrophase will correspond with the plots. I've updated the tests accordingly.