Update simulate_cosinor and its vignette
Updated the vignette to be clearer and more comprehensive
Previously, if there was a standard deviation specified for simulate_cosinor(), this standard deviation would be applied to both levels in a grouped-dataset. Now, the user can specify different standard deviations of 'alpha' parameters for a Gamma distribution. See the vignette for more details